javablog/src/main/java/models/Article.java
package models;

public class Article {

	private int id;
	private String title;
	private String imageLink;
	private String textContent;

	public Article(String title, String textContent) {
		this.title = title;
		this.textContent = textContent;
	}

	public int getId() {
		return id;
	}

	public void setId(int id) {
		this.id = id;
	}

	public String getTitle() {
		return title;
	}

	public void setTitle(String title) {
		this.title = title;
	}

	public String getImageLink() {
		return imageLink;
	}

	public void setImageLink(String imageLink) {
		this.imageLink = imageLink;
	}

	public String getTextContent() {
		return textContent;
	}

	public void setTextContent(String textContent) {
		this.textContent = textContent;
	}
}
